refactor(auth): stop re-exporting ConsentService from auth barrel
