+++ extension/src/extension/services/auth/index.ts
@@ -1,3 +1,6 @@
 export { AuthFlowHandler } from './authFlowHandler';
 export { AuthManager } from './authManager';
-export { ConsentService } from './consentService';
+// NOTE: ConsentService is intentionally NOT re-exported here. It must be imported
+// directly from './consentService' so the Open VSX (clean) build, which only pulls
+// AuthManager/AuthFlowHandler from this barrel, never drags consentService.ts into
+// the bundle. Import it via '@extension/services/auth/consentService'.
